+++ kadai1/shimastripe/cmd/kadai1/main.go
@@ -3,10 +3,10 @@ package main
 import (
 	"os"
 
-	"github.com/gopherdojo/dojo3/kadai1"
+	"github.com/gopherdojo/dojo3/kadai1/shimastripe"
 )
 
 func main() {
-	cli := &kadai1.CLI{OutStream: os.Stdout, ErrStream: os.Stderr}
+	cli := &shimastripe.CLI{OutStream: os.Stdout, ErrStream: os.Stderr}
 	os.Exit(cli.Run(os.Args))
 }
