+++ kadai1/shimastripe/traverse.go
@@ -1,4 +1,4 @@
-package kadai1
+package shimastripe
 
 import (
 	"os"
